requirements.txt
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
pillow==10.4.0
matplotlib==3.9.2
streamlit==1.38.0
tqdm==4.66.5
pyyaml==6.0.2
pytorch-grad-cam==1.5.2
opencv-python==4.10.0.84